consolidate CalibTracker/SiStripQuality
