+++ CalibTracker/SiStripQuality/plugins/SiStripQualityStatistics.cc
@@ -16,15 +16,59 @@
 //
 //
 
-#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+// system include files
+#include <memory>
+
+// user include files
+#include "CalibFormats/SiStripObjects/interface/SiStripQuality.h"
 #include "CalibTracker/SiStripCommon/interface/SiStripDetInfoFileReader.h"
+#include "CalibTracker/SiStripQuality/interface/SiStripQualityWithFromFedErrorsHelper.h"
 #include "CommonTools/TrackerMap/interface/TrackerMap.h"
+#include "DQM/SiStripCommon/interface/TkHistoMap.h"
+#include "DQMServices/Core/interface/DQMEDHarvester.h"
 #include "DataFormats/SiStripDetId/interface/StripSubdetector.h"
 #include "DataFormats/TrackerCommon/interface/TrackerTopology.h"
-#include "DQM/SiStripCommon/interface/TkHistoMap.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+#include "Geometry/Records/interface/TrackerTopologyRcd.h"
 
-#include "CalibTracker/SiStripQuality/plugins/SiStripQualityStatistics.h"
+class SiStripFedCabling;
+
+class SiStripQualityStatistics : public DQMEDHarvester {
+public:
+  explicit SiStripQualityStatistics(const edm::ParameterSet&);
+  ~SiStripQualityStatistics() override = default;
+
+  void endRun(edm::Run const&, edm::EventSetup const&) override;
+  void dqmEndJob(DQMStore::IBooker&, DQMStore::IGetter&) override;
+
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+
+private:
+  void updateAndSave(const SiStripQuality* siStripQuality);
+  void SetBadComponents(int, int, SiStripQuality::BadComponent&);
+
+  edm::RunID run_;
+  std::string dataLabel_;
+  const std::string TkMapFileName_;
+  const bool saveTkHistoMap_;
+  //Global Info
+  int NTkBadComponent[4];  //k: 0=BadModule, 1=BadFiber, 2=BadApv, 3=BadStrips
+  int NBadComponent[4][19][4];
+  //legend: NBadComponent[i][j][k]= SubSystem i, layer/disk/wheel j, BadModule/Fiber/Apv k
+  //     i: 0=TIB, 1=TID, 2=TOB, 3=TEC
+  //     k: 0=BadModule, 1=BadFiber, 2=BadApv, 3=BadStrips
+  std::stringstream ssV[4][19];
+
+  TrackerMap *tkMap, *tkMapFullIOVs;
+  SiStripDetInfo detInfo_;
+  std::unique_ptr<TkHistoMap> tkhisto;
+  const edm::ESGetToken<TrackerTopology, TrackerTopologyRcd> tTopoToken_;
+  const edm::ESGetToken<TkDetMap, TrackerTopologyRcd> tkDetMapToken_;
+  std::unique_ptr<TrackerTopology> tTopo_;
+  SiStripQualityWithFromFedErrorsHelper withFedErrHelper_;
+};
 
 SiStripQualityStatistics::SiStripQualityStatistics(const edm::ParameterSet& iConfig)
     : TkMapFileName_(iConfig.getUntrackedParameter<std::string>("TkMapFileName", "")),
@@ -121,7 +165,7 @@ void SiStripQualityStatistics::updateAndSave(const SiStripQuality* siStripQualit
                             ((BC[i].BadApvs >> 2) & 0x1) + ((BC[i].BadApvs >> 1) & 0x1) + ((BC[i].BadApvs) & 0x1);
 
     //&&&&&&&&&&&&&&&&&
-    //Single SubSyste
+    //Single SubSystem
     //&&&&&&&&&&&&&&&&&
     int component;
     DetId detectorId = DetId(BC[i].detid);
@@ -325,3 +369,6 @@ void SiStripQualityStatistics::SetBadComponents(int i, int component, SiStripQua
     tkMap->fillc(BC.detid, 0x0);
   }
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(SiStripQualityStatistics);
